requirements.txt
Flask==1.1.1
requests==2.22.0
gunicorn==20.0.4
google-cloud-vision==0.42.0
webexteamssdk==1.1.1